+++ energy_sequential.c
@@ -13,14 +13,14 @@ int main(int argc, char* argv[]) {
 
     double omega = strtod(argv[1], NULL);
     double t = strtod(argv[2], NULL);
-    long n = strtol(argv[3], NULL, 10);
+    long long n = strtol(argv[3], NULL, 10);
 
     double coeff = -(hbar * omega) / (k * t);
 
     double topSum = 0.0;
     double bottomSum = 0.0;
 
-    for (long count = 0; count <= n; count++) {
+    for (long long count = 0; count <= n; count++) {
         double temp = exp(coeff * count);
         topSum += (2*count + 1) * temp;
         bottomSum += temp;
